Refactors command handling to use sessions
Moves command execution logic into a session-based model.
This change replaces the direct passing of the database context
to commands with a session object that encapsulates the context
and connection-specific data. This allows for better management
of connection state and simplifies command execution. It also removes
the parsing of RESP commands from the server to the session level.
Removes the test files as those are related to the old database
context.