Refactors to RESP protocol and enhances commands (#9)
* Removes gRPC dependencies

Removes gRPC and protobuf related code.

The codebase is restructured and gRPC was deemed unnecessary.
Communication will now be handled over TCP.

* Implements basic TCP server

Sets up a basic TCP server that listens for incoming connections.
Handles client connections in separate goroutines.
Implements PING and SET commands.
Includes signal handling for graceful shutdown.

* Implements basic RESP protocol

Adds basic RESP protocol parsing and command execution.

Introduces:
- RESP parsing logic to handle client requests.
- Command interface for handling different commands.
- Ping command implementation.
- Error handling for unknown commands and invalid requests.

This commit sets the foundation for building more complex
commands and features in the future.

* Implements the HELLO command

Adds support for the HELLO command, enabling server discovery and protocol negotiation.

This change introduces the HELLO command handler and related error and result formatting utilities, ensuring compatibility with the RESP protocol. It also updates the command creation logic to include the HELLO command.

* Implements SET and GET commands

Adds implementation for the SET and GET commands, allowing users to store and retrieve data.

* Refactors command handling to use sessions

Moves command execution logic into a session-based model.
This change replaces the direct passing of the database context
to commands with a session object that encapsulates the context
and connection-specific data. This allows for better management
of connection state and simplifies command execution. It also removes
the parsing of RESP commands from the server to the session level.
Removes the test files as those are related to the old database
context.

* Include proper server version in HELLO command

* Adds DEL command to delete keys

Implements the DEL command, enabling the removal of one or more keys from the database.
The command returns the number of successfully deleted keys.
Handles database access and error checking for key deletion.

* Simplifies database access in session

Refactors session to directly provide a database instance, removing the need for repeated database retrieval within command invocations.

This streamlines database interactions and improves code readability.
Removes the now-unnecessary data, db, and sys server files.

* SELECT command (wip)

* Fully implementation of SELECT Command

* Test server with redis Go client

* Adds session name functionality

Implements the ability to set a name for a session using the HELLO command.

Adds a SETNAME option to the HELLO command to allow clients to associate a name with their session.

Updates the Session interface and implementation to include a SetName method.

* Adds CLIENT command with SETNAME/GETNAME

Adds the CLIENT command to allow setting and getting the client connection name.

This allows identifying connections for debugging and management purposes.

* Sync when creating new database

* There's no need to have DbContext

* Fix tests

* Fix logger in db

* Using singleton logger

* Change configuration from json to env

* Update README file